requirements.txt
Django
Pillow
djangorestframework
drf-nested-routers